Check for ERRORARRAY before checking its length
If the API client is called with a batch of commands, then
obj.ERRORARRAY does not exist since obj is an array.  We could
deal with the whole batching stuff in linode-api, or just let the
caller deal with it.

In either case, we should be prepared for the case where ERRORARRAY
does not actually exist, so checking its length throws an exception.