Return object if obj.DATA not set
This is also related to using the batch API.  If a caller uses the
batch API, then obj.DATA will not be set since obj is an array.  In
this case, we just pass obj back and let the caller do what they like.

There's probably a better way to deal with this, but this is the
simplest.